// libs/juridico/src/lib/models/SeccionDeterminativa/detCredImpuestoRtaModel.ts
export interface DetCredImpuestoRtaModel {
    refTabla: string;
    numFormul: string;
    mtoCas123: number;
    mtoCas136: number;
    mtoCas134: number;
    mtoCas126: number;
    mtoCas125: number;
    mtoCas504: number;
    mtoCas127: number;
    mtoCas128: number;
    mtoCas130: number;
    mtoCas129: number;
    mtoCas506: number;
    mtoCas131: number;
    mtoCas279: number;
    mtoCas202: number;
    mtoCas203: number;
    mtoCas204: number;
    mtoCas297: number;
    mtoCas783: number;
}

export interface DetCredImpuestoRtaModelCas126 {
    numFormul: string;
    desValLiteral: string;
    mtoLiteral: number;
}

export interface DetCredImpuestoRtaModelCas128 {
    numFormul: string;
    codTipMto: string;
    desValLiteral: string;
    mtoLiteral: number;
}

export interface DetCredImpuestoRtaModelCas130 {
    numFormul: string;
    desValLiteral: string;
    mtoLiteral: number;
}

export interface DetCredImpuestoRtaModelCas131 {
    numFormul: string;
    desValLiteral: string;
    mtoLiteral: number;
}

export interface DetCredImpuestoRtaModelCas297 {
    numFormul: string;
    numPerAplica: string;
    numPerPagoCta: string;
    numVal: string;
    codTri: string;
    mtoImpuesto: number;
    codOrigen: string;
    indActivo: string;
    numDoc: string;
    codFor: string;
    numRes: string;
}

export interface DeudasActivasModel {
    numFormul: string;
    numPerAplica: string;
    numPerPagoCta: string;
    numVal: string;
    codTri: string;
    mtoImpuesto: number;
    codOrigen: string;
    indActivo: string;
    numDoc: string;
    codFor: string;
    numRes: string;
    id: number;
}