// libs/juridico/src/lib/models/SeccionDeterminativa/impRtaEmpresaModel.ts
export interface ImpRtaEmpresaModel {
  refTabla: string;
  numFormul: string;
  mtoCas100: number;
  mtoCas101: number;
  mtoCas103: number;
  mtoCas105: number;
  mtoCas106: number;
  mtoCas107: number;
  mtoCas120: number;
  mtoCas108: number;
  mtoCas110: number;
  mtoCas113: number;
  mtoCas111: number;
  mtoCas686: number;
  mtoCas610: number;
  mtoCas880: number;
}